Create lib for public consumption
